Marshall data serialization
The marshall structure takes any data pointer and predicts its
scalar. Data is converted according to this type and stored in
the universal data structure. Serialization can rework the entire
universal data structure into a character array, and return it as
an string. Further core operations will all convert to or from the
universal datatype known as marsall_t.